Add PR review changes

+++ application/common/dos_location.py
@@ -7,8 +7,8 @@ class DoSLocation:
 
     id: int
     postcode: str
-    easting: int
-    northing: int
+    easting: float
+    northing: float
     postaltown: str
     latitude: float
     longitude: float
